apps/windows/python/Whisper/requirements.txt
opencv-python
qai_hub_models[whisper_base_en]
audio2numpy
samplerate
onnxruntime
onnxruntime-qnn
